packet interface refactoring
- Service codes are chosen by user programs.  Theretically a program may
  discover up to 32768 services.

- Service discovery and pollout packets have their own built-in codes.

- Services may be discovered incrementally.

- Flags API types in have unspecified sizes.

- Internal trap packet is identified by a code instead of a flag.

Other changes:

- gate_service_registry API already supports asynchronous service info
  change notifications (but runtime doesn't support that yet).

- ServiceRegistry supports only one instance of a given service
  implementation per program instance.  Secondary discoveries are flagged
  as unavailable.

- ServiceRegistry.Clone() forms a hierarchy instead of copying data.

+++ packet/packet.go
@@ -16,9 +16,9 @@ const (
 // Code represents the source/destination of a packet.
 type Code [2]byte
 
-// Int is mostly useful for debug logging.
-func (code Code) Int() int {
-	return int(binary.LittleEndian.Uint16(code[:]))
+// Int16 is mostly useful for debug purposes.
+func (code Code) Int16() int16 {
+	return int16(binary.LittleEndian.Uint16(code[:]))
 }
 
 // BufSize calculates packet size based on packet content size.
